Add a fast path for unwrapping if in AGG IF to FILTER rewrite
The optimization of unwrapping if has a large impact on the query
performance. For example, here is the improvement of this optimization
for one specific prod query:
Original CPU with this rule disabled: 1.79h
CPU with this rule enabled without unwrapping IF: 1.42h
CPU with this rule enabled with unwrapping IF: 17m

Thus it seems worth adding a fast path for the IF unwrapping.
This commit adds a new session property to enable/disable this feature.
This also adds the logic to disable IF unwrapping for some query
patterns with known issues, e.g., array[1] and a/b which could return
exceptions. We can add more query patterns to the block list if
necessary.

We are not planning to enable this fast path by default. It will only be
enabled for specific clusters like Deltoid which don't have query
patterns with these issues.