Unwrap IF for more cases in AGG IF to FILTER rewrite
Previous blocklist disables the IF unwrapping based on the function type
in the true branch expression. E.g., if there is a division or
array offset function, we disable the IF unwrapping. This blocklist is
too aggressive. There are many prod queries with divisions cannot
benefit from this rewrite.

Actually, if the condition expression doesn't have any variables
referenced in the true branch expression, we can safely unwrap the IF
expression. This commit adds this check to allow enabling the rewrite
for more queries safely.